Lockscreen custom wallpaper (1/2)
kitkat now has code for a custom wallpaper image.

Using that code set a switch case to set the custom background between
	- Color Fill
	- Custom Wallpaper
	- Default background

This way allows the same overlay shadow thats set on the keyguard when its
the defualt way. And the overlay effect works on both color fill and
custom image.

Change-Id: I98fde63ef9f3c0a5f334cd9a97b237a6f99c57f6